comment_introduction entity create
