+++ src/main/java/com/rocket/rocket_project/recruit/entity/SkillRecruit.java
@@ -1,7 +1,7 @@
-package com.rocket.rocket_project.skill_recruit.entity;
+package com.rocket.rocket_project.recruit.entity;
 
 import com.rocket.rocket_project.recruit.entity.Recruit;
-import com.rocket.rocket_project.skill.entity.Skill;
+import com.rocket.rocket_project.recruit.entity.Skill;
 import lombok.*;
 import javax.persistence.*;
 import java.io.Serializable;
